+++ packages/debug_kit/lib/src/ui/overlay/debug_kit_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../../core/controller/debug_kit_controller.dart';
-import '../screens/debug_kit_console_screen.dart';
+import 'debug_kit_console_launcher.dart';
 
 /// The draggable floating debug button shown by [DebugKitOverlay].
 class DebugKitButton extends StatefulWidget {
@@ -44,31 +44,6 @@ class _DebugKitButtonState extends State<DebugKitButton> {
     );
   }
 
-  void _openConsole(BuildContext context) {
-    NavigatorState? navigator = Navigator.maybeOf(context);
-
-    if (navigator == null) {
-      final config = DebugKitController().config;
-      if (config.navigatorKey != null) {
-        navigator = config.navigatorKey!.currentState;
-      }
-    }
-
-    if (navigator != null) {
-      navigator.push(
-        MaterialPageRoute(
-          builder: (_) => const DebugKitConsoleScreen(),
-          settings: const RouteSettings(name: 'debug_kit_console'),
-        ),
-      );
-    } else {
-      // ignore: avoid_print
-      print(
-          'DebugKit: Could not find Navigator. Ensure you are calling this from a context '
-          'descended from Navigator or provide a navigatorKey during DebugKit.init().');
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     if (_position == null) return const SizedBox.shrink();
@@ -93,7 +68,7 @@ class _DebugKitButtonState extends State<DebugKitButton> {
                 _clampPosition(screenSize);
               });
             },
-            onTap: () => _openConsole(context),
+            onTap: () => openDebugKitConsole(context: context),
             child: SizedBox(
               width: _buttonSize + 10, // Extra touch area
               height: _buttonSize + 10,